naemon.c: do not reset shutdown and restart flags early
Previously we reset these flags in the top of the main loop
meaning that if we were restarting the process we could never
detect during the second startup if we were actually restarting.

Instead the state would be that a restart would equal a fresh
bootup of the process.

Part of MON-10376

Signed-off-by: Robin Engström <robin.engstrom@op5.com>