Check snprintf for encoding errors and truncation
GCC 7 introduced warning messages for possible truncation of output from
snprintf (and friends). This introduced new warnings when building
Naemon with GCC 7. As we compile with -Werror this break the build.

This patch first checks for any encoding errors that might error as a
result of snprintf, and thereafter ensures that the output was not
truncated. If any errors occurs a warning message is printed.

In checks.c the call to snprintf was moved inside an if statement as
the result of snprintf is only used inside the mentioned if statement.
This ensures we only print warnings if the file is actually used.

This is the second of two patches to fix
https://github.com/naemon/naemon-core/issues/184 / MON-10862

Signed-off-by: Jacob Hansen <jhansen@op5.com>