Notification is sent even if a service is in a scheduled flexible downtime
When a service is set in flexible downtime, the first notification is sent out before the downtime schedule takes effect.
Since service result check logic is just add the downtime start event to the event queue and then start notification right after,
this bug is natural with this flow.
The solution is checking whether there is any flexible downtime started and if not we will send out the normal notifciation.

Link: https://jira.op5.com/browse/MON-8986

Signed-off-by: Nian Tang <ntang@op5.com>