Triggered Scheduled downtime
The triggered scheduled downtime is deleted when the triggering downtime is starting.
This issue is due to that flex downtime start variable is feeded with right value.
It will add a wrong time (negative) for the downtime stop event on the event queue and
the downtime stop event for the triggered downtime will start evecution right after it is started.
The solution is to initialize the flex downtime start variable of the triggered downtime with
the value of the flex downtime start value of the triggering one.

https://jira.op5.com/browse/MON-8982

Signed-off-by: Nian Tang <ntang@op5.com>