Switching service states for host down is confusing and ruins reports
When there is a service check failure and the host is down, the service will be set to hard state
even though the service has not reach the max attempts. This behavior is corrected and notification is not send
while the host is down as it is supposed to be.

Signed-off-by: Nian Tang <ntang@op5.com>