Double naemon process
Double naemon process situtation is created if other process is monitoring naemon process and try to restart it.
Naemon relies on the pid file locking to ensure other instance of naemon could not be started. But the advisory locking breaks if the
pid file is remove. Bascially you can always create a new naemon process if you just remove the pidfile.
In the commit, the pidfile is not removed after naemon process exiting the main function and the modified the stop function in the
init.d file so that the pidfile is not removed after killing the naemon process

Signed-off-by: Nian Tang <ntang@op5.com>