No on-demand host checks when service is in hard critical
The current Naemon documentation states that on-demand host checks are
performed when a service associated with the host changes state.
This is not actually what happens, currently on-demand host checks
are scheduled on any service checks which results in a NON-OK state for
the service.

This commit changes that logic slightly, so that on-demand host checks
are performed when the service changes state OR if the service is in a
soft state. That is, we no longer do on-demand host checks on non-ok
service checks, if the service is already in hard critical.

This ensures that we do not do a ridiculous amount of host checks when a
host and all services are in hard critical.

This fixes MON-5625

Signed-off-by: Jacob Hansen <jhansen@op5.com>