Init: Increase time till SIGKILL is sent
When using killproc to kill a process, first a SIGTERM signal is sent,
and after a default of 3 seconds (5 seconds on suse), a SIGKILL is sent.

On larger setups we often see that 3 seconds, is not sufficient for
Naemon to shutdown, and as a result retention data might not be
correctly saved.

This commit increase the timeout to 90 seconds on rehat based releases
and on SUSE. This micics the behaviour of systemd.

This is part of MON-10565

Signed-off-by: Jacob Hansen <jhansen@op5.com>