Retain next_check schedule on restart
This commit ensures that the next_check schedule for hosts and services
are retained on Naemon restart, given that use_retained_scheduling_info
is enabled.

The logic is as follows:

- If use_retained_scheduling_info is disabled, set a random time (as
  before)
- If use_retained_schedule_info is enabled:
  - If we didn't miss the check during the restart, retain the old
    next_check time
  - If we missed one check, schedule the service/host within the next
    interval_length (usually 60 seconds)
  - If we missed more than one check, schedule the next check randomly.

We schedule missed checks within 60 seconds, rather than immediately in
order to do some load balacing. This is also the rationale for
scheduling the check randomly, in case we missed more than one check
(this indicates Naemon has been down for a longer period of time).

This fixes:
- https://github.com/naemon/naemon-core/issues/224
- https://github.com/naemon/naemon-core/issues/156
- MON-10720 (https://jira.op5.com/browse/MON-10720)

Signed-off-by: Jacob Hansen <jhansen@op5.com>