Fix heap corruption when callback dereigsters itself
A neb callback deregistering itself in a callback currently causes a heap
corruption. This is due to getting the naming information for the module
after the callback has been run.

This patch gets the naming information of the module before the
callback, ensuring that we will not try to access the callback pointer
after it is potentially freed.

This fixes: MON-11365 & https://github.com/naemon/naemon-core/issues/268

Signed-off-by: Jacob Hansen <jhansen@op5.com>