Don't process perfdata file if perfdata is disabled
Currently, even if `process_performance_data` is set to 0, the commands
set by `host/service_perfdata_file_processing_command` are executed.
Often these file processing commands move the current perfdata files to
a spool directory. These files will be empty if
`process_performance_data` is set to 0. Given the files are empty
anyway, it does not make a lot of sense to process the perfdata files.
Furthermore, if no software is cleaning up the empty files from the spool
directory, the system will eventually be flooded with empty files.

This commit fixes the above behaviour by not executing the file
processing commands if performance data processing is disabled.

Signed-off-by: Jacob Hansen <jhansen@op5.com>