Downtimes: don't send NEB callback if add failed
Today, even if we fail to add a host or service downtime we still send a
NEB callback. This results in Livestatus adding a downtime which as
never actually added in Naemon.

This commit ensures we only send the NEB callback with
`NEBTYPE_DOWNTIME_ADD` if we succeeded in adding the downtime.

Signed-off-by: Jacob Hansen <jhansen@itrsgroup.com>