add dummy filter hostname for each host in hostgroup
delete_downtime_by_hostname_service_description_start_time_comment()
requires at least one of the filter parameters  to be set.  in our case,
it should be possible to delete *all* downtime for a given hostgroup, so
it is valid to not provide any filters.  luckily, delete_if_matches is
called for each host in hostgroup, and the hostname is passed as _name
_name is actually the key value for each host in hostgroup.members,
and this is a normal string containing the host's identifying name.

we can therefore pass this on as a dummy filter.