Copy timeperiod names into contact struct members
This fixes a problem where we'd copy the address instead of the content
of the timeperiod name to the host_notification_period and
service_notification_period members of contacts, causing a crash later
when they were free'd.

t-tap/test_commands now passes again.

Found while looking for the cause of #285, but doesn't seem to be
related.

Signed-off-by: Roel van Meer <roel@1afa.com>