add problem timestamps and duration
this PR makes hosts / services save the start and end timestamp of the current
problem. Those values can then be used as macros, ex. in notification scripts.
For this, there are several new macros available:

- $HOSTPROBLEMSTART$ start timestamp of problem
- $HOSTPROBLEMEND$ end timestamp of problem (or zero if problem still persists)
- $HOSTPROBLEMDURATIONSEC$ duration of problem
- $HOSTPROBLEMDURATION$ duration as human readable text

the same macros exist for services:
- $SERVICEPROBLEMSTART$
- $SERVICEPROBLEMEND$
- $SERVICEPROBLEMDURATIONSEC$
- $SERVICEPROBLEMDURATION$

While there is a currently ongoing problem, the values point to this current
problem and the end timestamp is zero. Once the problem is resolved, the values
can still be used and won't change until a new problem starts.

This makes it possible to use the problem duration in recovery notifications which
otherwise would not be possible.