add new options to keep services running as long as they are up
The issue with the options:

- host_down_disable_service_checks
- service_skip_check_dependency_status
- service_skip_check_host_down_status
- host_skip_check_dependency_status

is that reports break because hosts/services suddenly stop executing and
keep their OK state. Which makes those options pretty unusable.

So in order to keep reporting correct, you need to keep services running, even
if the host is down. With these new options, hosts/services keep on running as
long as they are up. And as soon as the service is down, it stops running until
the host comes back up. That way naemon has to do less checks, especially less
checks which run into timeouts and such but reporting is still correct.

The option service_skip_check_dependency_status=-2 will also be used for service parents.

Adding a new option service_parents_disable_service_checks to prevent running service
checks if service parents are down.

Recommended settings are:

    host_down_disable_service_checks=1       ; disable service checks if host is down
    service_parents_disable_service_checks=1 ; also disable service checks if parents are down
    service_skip_check_host_down_status=-2   ; but keep running as long as they are ok
    service_skip_check_dependency_status=-2  ; same, but for dependency checks.
    host_skip_check_dependency_status=-2     ; and for host checks.