fix handling performance data
the spool file way of handling performance data works like this:

- check result comes in
- perf data gets extracted
- write perf data into spool file
- close spool file
- call spool file handler
- open spool file again

so during the runtime of the spool handler the spool file is closed for good reasons. But the check result handler
simply skips adding performance data when the spool file is closed which basically means during about a second
we loose all performance data.
So instead of checking the spool file, we better check the buffer queue which is used to buffer the
performance data already.

Signed-off-by: Sven Nierlein <sven@consol.de>