fix pending service dependencies
Having a new host with all services pending 'service_skip_check_dependency_status' does not work as expected.
The idea was to keep services running until they fail to not break reporting. But only if they ever worked
before. So we need to check the 'has_been_checked' flag as well.